Moved test to bottomtabs tests
